requirements.txt
flask
pillow
cryptography
gunicorn